Move unnamed requirements to their own pep508_rs module and requirements-txt
Another refactoring in preparation of using a richer requirements type. No functional changes, only moves code around
